# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "National Exam Card API",
        "description": "The National Exam Card API provides access to student's application for exam card data and process. ",
        "version": "0.1.0"
    },
    "paths": {
...
        "/calendar": {
            "get": {
                "tags": [
                    "calendar"
                ],
                "summary": "Get a list of all calendar events",
                "operationId": "08a803fe237a76081b119fcd0961cf09",
                "responses": {
                    "200": {
                        "description": "List of calendar events"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "calendar"
                ],
                "summary": "Create a new calendar event",
                "operationId": "857913d662da8c9a5e14b0dd3ebc109d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sample Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Sample content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Calendar event created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/calendar/{id}": {
            "get": {
                "tags": [
                    "calendar"
                ],
                "summary": "Get a specific calendar event by ID",
                "operationId": "8b1a3cb09aaf8b498b22e612257af49d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the calendar event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Calendar event details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "calendar"
                ],
                "summary": "Update an existing calendar event",
                "operationId": "c9feb19dc9e866e9c6963ef1808c64ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the calendar event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calendar event updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "calendar"
                ],
                "summary": "Delete a calendar event by ID",
                "operationId": "8eed73fd20b5ba6a5d83dc26d9f70f2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the calendar event",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Calendar event deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}